+++ packages/compiler-cli/src/ngcc/test/analysis/switch_marker_analyzer_spec.ts
@@ -8,7 +8,7 @@
 import * as ts from 'typescript';
 
 import {SwitchMarkerAnalyzer} from '../../src/analysis/switch_marker_analyzer';
-import {Fesm2015ReflectionHost} from '../../src/host/fesm2015_host';
+import {Esm2015ReflectionHost} from '../../src/host/esm2015_host';
 import {makeProgram} from '../helpers/utils';
 
 const TEST_PROGRAM = [
@@ -47,7 +47,7 @@ describe('SwitchMarkerAnalyzer', () => {
   describe('analyzeProgram()', () => {
     it('should check for switchable markers in all the files of the program', () => {
       const program = makeProgram(...TEST_PROGRAM);
-      const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+      const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
       const analyzer = new SwitchMarkerAnalyzer(host);
       const analysis = analyzer.analyzeProgram(program);
