+++ packages/compiler-cli/src/ngcc/test/host/esm2015_host_import_helper_spec.ts
@@ -9,7 +9,7 @@
 import * as ts from 'typescript';
 
 import {ClassMemberKind, Import} from '../../../ngtsc/host';
-import {Fesm2015ReflectionHost} from '../../src/host/fesm2015_host';
+import {Esm2015ReflectionHost} from '../../src/host/esm2015_host';
 import {convertToDirectTsLibImport, getDeclaration, makeProgram} from '../helpers/utils';
 
 const FILES = [
@@ -104,7 +104,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
       describe('getDecoratorsOfDeclaration()', () => {
         it('should find the decorators on a class', () => {
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
           const decorators = host.getDecoratorsOfDeclaration(classNode) !;
@@ -121,14 +121,14 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
         });
 
         it('should use `getImportOfIdentifier()` to retrieve import info', () => {
-          const spy = spyOn(Fesm2015ReflectionHost.prototype, 'getImportOfIdentifier')
+          const spy = spyOn(Esm2015ReflectionHost.prototype, 'getImportOfIdentifier')
                           .and.callFake(
                               (identifier: ts.Identifier) => identifier.getText() === 'Directive' ?
                                   {from: '@angular/core', name: 'Directive'} :
                                   {});
 
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
 
@@ -143,7 +143,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
 
         it('should support decorators being used inside @angular/core', () => {
           const program = makeProgram(fileSystem.files[1]);
-          const host = new Fesm2015ReflectionHost(true, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(true, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/node_modules/@angular/core/some_directive.js', 'SomeDirective',
               ts.isVariableDeclaration);
@@ -164,7 +164,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
       describe('getMembersOfClass()', () => {
         it('should find decorated members on a class', () => {
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
           const members = host.getMembersOfClass(classNode);
@@ -182,7 +182,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
 
         it('should find non decorated properties on a class', () => {
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
           const members = host.getMembersOfClass(classNode);
@@ -196,7 +196,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
 
         it('should find static methods on a class', () => {
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
           const members = host.getMembersOfClass(classNode);
@@ -209,7 +209,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
 
         it('should find static properties on a class', () => {
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
 
@@ -223,10 +223,10 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
 
         it('should use `getImportOfIdentifier()` to retrieve import info', () => {
           const spy =
-              spyOn(Fesm2015ReflectionHost.prototype, 'getImportOfIdentifier').and.returnValue({});
+              spyOn(Esm2015ReflectionHost.prototype, 'getImportOfIdentifier').and.returnValue({});
 
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
 
@@ -237,7 +237,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
 
         it('should support decorators being used inside @angular/core', () => {
           const program = makeProgram(fileSystem.files[1]);
-          const host = new Fesm2015ReflectionHost(true, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(true, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/node_modules/@angular/core/some_directive.js', 'SomeDirective',
               ts.isVariableDeclaration);
@@ -253,7 +253,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
       describe('getConstructorParameters', () => {
         it('should find the decorated constructor parameters', () => {
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
           const parameters = host.getConstructorParameters(classNode);
@@ -270,11 +270,11 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
         describe('(returned parameters `decorators`)', () => {
           it('should use `getImportOfIdentifier()` to retrieve import info', () => {
             const mockImportInfo = {} as Import;
-            const spy = spyOn(Fesm2015ReflectionHost.prototype, 'getImportOfIdentifier')
+            const spy = spyOn(Esm2015ReflectionHost.prototype, 'getImportOfIdentifier')
                             .and.returnValue(mockImportInfo);
 
             const program = makeProgram(fileSystem.files[0]);
-            const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+            const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
             const classNode = getDeclaration(
                 program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
             const parameters = host.getConstructorParameters(classNode);
@@ -292,7 +292,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
       describe('getDeclarationOfIdentifier', () => {
         it('should return the declaration of a locally defined identifier', () => {
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
           const ctrDecorators = host.getConstructorParameters(classNode) !;
@@ -308,7 +308,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
 
         it('should return the declaration of an externally defined identifier', () => {
           const program = makeProgram(fileSystem.files[0]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const classNode = getDeclaration(
               program, '/some_directive.js', 'SomeDirective', ts.isVariableDeclaration);
           const classDecorators = host.getDecoratorsOfDeclaration(classNode) !;
@@ -331,7 +331,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
       describe('getVariableValue', () => {
         it('should find the "actual" declaration of an aliased variable identifier', () => {
           const program = makeProgram(fileSystem.files[2]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const ngModuleRef = findVariableDeclaration(
               program.getSourceFile(fileSystem.files[2].name) !, 'HttpClientXsrfModule_1');
 
@@ -346,7 +346,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
 
         it('should return null if the variable has no assignment', () => {
           const program = makeProgram(fileSystem.files[2]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const missingValue = findVariableDeclaration(
               program.getSourceFile(fileSystem.files[2].name) !, 'missingValue');
           const value = host.getVariableValue(missingValue !);
@@ -355,7 +355,7 @@ describe('Fesm2015ReflectionHost [import helper style]', () => {
 
         it('should return null if the variable is not assigned from a call to __decorate', () => {
           const program = makeProgram(fileSystem.files[2]);
-          const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+          const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
           const nonDecoratedVar = findVariableDeclaration(
               program.getSourceFile(fileSystem.files[2].name) !, 'nonDecoratedVar');
           const value = host.getVariableValue(nonDecoratedVar !);
