+++ packages/compiler-cli/src/ngcc/test/rendering/esm2015_renderer_spec.ts
@@ -13,7 +13,7 @@ import {makeProgram} from '../helpers/utils';
 import {DecorationAnalyzer} from '../../src/analysis/decoration_analyzer';
 import {SwitchMarkerAnalyzer} from '../../src/analysis/switch_marker_analyzer';
 import {DtsMapper} from '../../src/host/dts_mapper';
-import {Fesm2015ReflectionHost} from '../../src/host/fesm2015_host';
+import {Esm2015ReflectionHost} from '../../src/host/esm2015_host';
 import {Esm2015Renderer} from '../../src/rendering/esm2015_renderer';
 
 function setup(file: {name: string, contents: string}, transformDts: boolean = false) {
@@ -21,7 +21,7 @@ function setup(file: {name: string, contents: string}, transformDts: boolean = f
   const dtsMapper = new DtsMapper(dir, dir);
   const program = makeProgram(file);
   const sourceFile = program.getSourceFile(file.name) !;
-  const host = new Fesm2015ReflectionHost(false, program.getTypeChecker());
+  const host = new Esm2015ReflectionHost(false, program.getTypeChecker());
   const decorationAnalyses =
       new DecorationAnalyzer(program.getTypeChecker(), host, [''], false).analyzeProgram(program);
   const switchMarkerAnalyses = new SwitchMarkerAnalyzer(host).analyzeProgram(program);
